ast_results/VREMSoftwareDevelopment_WifiAnalyzer/app/src/test/java/com/vrem/wifianalyzer/wifi/timegraph/TimeGraphCacheTest.java
// isComment
package com.vrem.wifianalyzer.wifi.timegraph;

import com.vrem.wifianalyzer.wifi.band.WiFiWidth;
import com.vrem.wifianalyzer.wifi.graphutils.GraphConstants;
import com.vrem.wifianalyzer.wifi.model.WiFiDetail;
import com.vrem.wifianalyzer.wifi.model.WiFiSignal;
import org.apache.commons.collections4.Closure;
import org.apache.commons.collections4.IterableUtils;
import org.apache.commons.lang3.StringUtils;
import org.junit.Before;
import org.junit.Test;
import java.util.ArrayList;
import java.util.List;
import java.util.Set;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

public class isClassOrIsInterface {

    private TimeGraphCache isVariable;

    @Before
    public void isMethod() {
        isNameExpr = new TimeGraphCache();
    }

    @Test
    public void isMethod() {
        // isComment
        List<WiFiDetail> isVariable = isMethod();
        // isComment
        Set<WiFiDetail> isVariable = isNameExpr.isMethod();
        // isComment
        isMethod(isNameExpr.isMethod(), isNameExpr.isMethod());
    }

    @Test
    public void isMethod() {
        // isComment
        List<WiFiDetail> isVariable = isMethod();
        // isComment
        Set<WiFiDetail> isVariable = isNameExpr.isMethod();
        // isComment
        isMethod(isNameExpr.isMethod() - isIntegerConstant, isNameExpr.isMethod());
        isMethod(isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant)));
    }

    @Test
    public void isMethod() {
        // isComment
        List<WiFiDetail> isVariable = isMethod();
        // isComment
        isNameExpr.isMethod();
        // isComment
        Set<WiFiDetail> isVariable = isNameExpr.isMethod();
        isMethod(isNameExpr.isMethod() - isIntegerConstant, isNameExpr.isMethod());
        isMethod(isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant)));
    }

    @Test
    public void isMethod() {
        // isComment
        List<WiFiDetail> isVariable = isMethod();
        // isComment
        isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant));
        // isComment
        Set<WiFiDetail> isVariable = isNameExpr.isMethod();
        isMethod(isNameExpr.isMethod(), isNameExpr.isMethod());
        isMethod(isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant)));
    }

    private WiFiDetail isMethod(String isParameter) {
        return new WiFiDetail(isNameExpr, "isStringConstant", isNameExpr.isFieldAccessExpr, new WiFiSignal(isIntegerConstant, isIntegerConstant, isNameExpr.isFieldAccessExpr, isIntegerConstant));
    }

    private List<WiFiDetail> isMethod() {
        List<WiFiDetail> isVariable = new ArrayList<>();
        for (int isVariable = isIntegerConstant; isNameExpr < isIntegerConstant; isNameExpr++) {
            WiFiDetail isVariable = isMethod("isStringConstant" + isNameExpr);
            isNameExpr.isMethod(isNameExpr);
        }
        isNameExpr.isMethod(isNameExpr, new AddClosure());
        for (int isVariable = isIntegerConstant; isNameExpr < isNameExpr.isFieldAccessExpr; isNameExpr++) {
            isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant));
        }
        return isNameExpr;
    }

    private class isClassOrIsInterface implements Closure<WiFiDetail> {

        @Override
        public void isMethod(WiFiDetail isParameter) {
            isNameExpr.isMethod(isNameExpr);
        }
    }
}
